bookstore_api/requirements.txt
aniso8601==3.0.0
Django==1.11.13
django-cors-headers==2.2.0
django-environ==0.4.4
django-filter==1.1.0
django-graphql-jwt==0.1.10
django-model-utils==3.1.2
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
graphene==2.1.1
graphene-django==2.0.0
graphql-core==2.0
graphql-relay==0.4.5
iso8601==0.1.12
Markdown==2.6.11
promise==2.1
PyJWT==1.6.4
pytz==2018.4
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
typing==3.6.4
